Made jQuery-UI work a little.
Have a couple of sample widgets included in Exercises index page;
a button and a datepicker. They work and serve as examples for
now. Also, have some style items from the default theme. Note
that ordinary tables and lists do not seem to be styled. My
css style on the main <ul> in the index page is being used.