Added some jQuery-UI theme features. Not finished.
Wanted to use jQuery-UI to allow various themes that eventually
the user would be allowd to change. Ran into trouble in that ordinary
text fields are not styled! I can't believe it. Besides text fields, a
couple of other simple fields are not styled either. I left it with
the default 'smoothness' theme connected via the applicaton.html.erb
file.