Added first lookup table: 'lookup_skills'. Now run rake db:seed .
The Exercises views now have a 'lookup_skill_id' foreign key column
to add the skill associated with the exercise. This required a
'lookup_skills' table with id as it's primary key and changes to
the model and views associated. Also changed the db/seeds.rb file
to contain initial values for this and future lookup tables.

The sort_order column indicates the proper order to list the
skills in. The application could sort by the 'skillset' column,
but their may be a signfican order within a skillset. Thus, the
need for the 'sort_order' column.