Changed the name of the sequencing column in lookup_skills table.
The column that indicated the proper sequence of records in
the lookup_skills column had been named 'sort_order'. This
didn't seem to communicate it's purpose well enough, so
George changed it to 'line_num' after the numbered lines
of an old BASIC program. This was to honor Dave's suggestion
that we number the lines by 10.