Got Twitter Bootstrap actually working.
The basic Bootstrap loads and finds the javascript and stylesheets
without error.

Have a main nav menu working across the top of all pages,
but the entries are mostly dummy placeholders. There is lots left
to do to get the styling and menuing in shape, but what's here is
working.