Add Twitter Bootstrap and user roles
This is a major commit. First Twitter bootstrap was added to provide
formatting.

Then user roles, via rolify gem, were added to give roles of
'teacher', 'parent' that the user can select from in his profile
page. This was implemented in the view for devise/registrations and
isn't really finished. For example, the 'admin' role was added, but
the user cannot self select into the 'admin' role. Instead, the
rolify command line is used at this time to create the administrator.

In time, more roles need to be added including the ability
for an administrator to add another administrator. Also, the
moderator role is needed to prune various exercises and users. See
http://railscasts.com/moderators for a discussion on what moderators
do.