Fixed Submit button and password accordion display panel
The Twitter Bootstrap submit button was styled as inactive, but
was still active. Changed it to be disabled / enabled.

The password and password confirmation fields are not shown by
default to make the user be more intentional about changing them.
However, if the submission is rejected for a reason associated
with these two fields, they need to be shown when the page is
re-rendered. This fix renders the password accordion panel for
most problems, but is not a total fix.

If the user enters a password, or both a password and a
password_confirmation, I can detect it and open the display
panel. However, if the user leaves password blank and submits only
a password_confirmation field, I cannot detect it as Rails
seems to blank this comfirmation field before my server-side Ruby
sees the value. So, in this case the view will not have the
passwords panel opened. However, this should be rare.