Fixed color on error messages when user edits profile.
The registrations page, like the user profile, allows a user
to change email, password, and roles. The standard devise
error messages had content, but were not styled. This is likely
because I changed the structure of the labels and input field and
the style did not find the fields. So, I updated the style for
this particular page.

Also updated the registrations controller to better test for
the existance of a password being changed. The change password
functionality is normally hidden, so this test is necessary to
open the hidden accordion panel to see the changed passwords.